Add simple username check.
I was HACKED!!!  Some user @MO_SO found my bot and opened the door and
then just left it open.  !!!???  The nerve.  It closed itself.

The user @MO_SO has a strange B&W picture of a woman with arabic writing
on it as their profile picture.  Strange.  Don't know how they found my
bot.  But it was inevitable, really.